### 2024-11-18 #7498 @kyleniemeyer
Hello @olgadoronina thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.

While we are waiting, could you please:

- Check the bibtex related error above? The paper compiled, but there seems to be an error. You can use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references:

- Suggest some potential reviewers from our database? Please do not actually tag them with @
