+++ paper/paper.bib
@@ -1,11 +1,3 @@
-%@book{aspen2022,
-%    title={Aspen plus: chemical engineering applications},
-%    author={Al-Malah, Kamal IM},
-%    year={2022},
-%    publisher={John Wiley \& Sons},
-%    doi = {10.1002/9781119293644}
-%}
-
 % Preatreatment
 @article{sitaraman_multiphysics_2015,
     title = {Multiphysics modeling and simulation of high-solids dilute-acid pretreatment of corn stover in a steam-explosion reactor},
