+++ paper/paper.bib
@@ -32,7 +32,8 @@
   volume={12},
   pages={1--15},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/s13068-019-1633-2}
 }
 
 % Bioreactor
@@ -126,7 +127,8 @@
   author={Humbird, D. and Davis, R. and Tao, L. and Kinchin, C. and Hsu, D. and Aden, A. and Schoen, P. and Lukas, J. and Olthof, B. and Worley, M, and Sexton, D. and Dungeon, D},
   year={2011},
   institution={National Renewable Energy Lab.(NREL), Golden, CO (United States)},
-  url={https://www.nrel.gov/docs/fy11osti/47764.pdf}
+  url={https://www.nrel.gov/docs/fy11osti/47764.pdf},
+  doi={10.2172/1013269}
 }
 
 @article{begum2024review,
