### 2024-11-19 #7498 @olgadoronina
Hello @kyleniemeyer,

thanks for managing the paper. We corrected the Bibtex file, and here is the list of possible reviewers for the paper (not from the reviewer database):
- Yidong Xia (INL), GitHub: yidongxiainl
- Olawufemi Oyedji (ORNL). Github: ooyedeji
- Mathew Liberatore (University of Toledo), email: matthew.liberatore@utoledo.edu
- Joe Saminiuk (Colorado School of Mines), email: samaniuk@mines.edu
