### 2025-03-06 #7498 @victoraalves
> Howdy! [@victoraalves](https://github.com/victoraalves), [@fwitte](https://github.com/fwitte), [@VEZY](https://github.com/VEZY), and [@DCoppitters](https://github.com/DCoppitters), would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.07498/joss.07498/10.21105.joss.07498.pdf))? The corresponding software is at this [link](https://github.com/NREL/VirtualEngineering).
>
> This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers almost everything about the review process; I will be happy to help with more information.
>
> Please let me know if you are interested. Thank you!

Hello! Yes, I am more than happy to review this one.
