+++ paper/paper_old.md
@@ -28,6 +28,10 @@ affiliations:
 date: 28 May 2024
 bibliography: paper.bib
 
+# # Optional fields if submitting to a AAS journal too, see this blog post:
+# # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
 # Summary
@@ -36,6 +40,7 @@ bibliography: paper.bib
 VE supports a wide variety of different multi-physics models and joins them to simulate an entire end-to-end process. To automate the execution of this model sequence, `VE` provides (i) a robust method to communicate between models, 
 (ii) a high-level, user-friendly interface to set model parameters and enable optimization, and (iii) an overall model-agnostic approach that allows new computational units to be swapped in and out of workflows. 
 Although the `VE` framework was developed to support the biochemical conversion of biomass to fuel, we have designed each component to easily accommodate new domains and unit models.
+<!-- Although the VE framework currently supports the biochemical conversion of biomass to fuel, we have designed each component to easily accommodate new domains and unit models. -->
 
 # Statement of need
 
@@ -43,7 +48,7 @@ Many industrial and manufacturing operations consist of a sequence of discrete p
 
 Numerical simulations that can support the analysis and optimization of such systems frequently require linking multiple individual models together—each associated with different steps in the overall operational process—so that the outputs of one model can inform the inputs of the next. These models can span multiple levels of physical fidelity and computational costs. `Virtual Engineering` (`VE`) is a Python package that enables the creation of this type of model sequence. 
 
-`VE` was originally developed to support the simulation and optimization of the biochemical conversion of lignocellulosic biomass to fuels. This bioconversion process was modeled by linking previously developed computational models as three important unit operations (Figure \ref{fig:VE_diagram}):  (i) the _pretreatment_ of the feedstock to make cellulose more accessible [@sitaraman_multiphysics_2015], (ii) an _enzymatic hydrolysis_ step to digest lignocellulose into sugars [@sitaraman_coupled_2019;@lischeske2019two], and (iii) a _bioconversion_ step to convert sugars into products in a bioreactor [@rahimi_computational_2018].  Finally, the capital and operating costs of the process, and the product's (in this case, ethanol) subsequent minimum feasible selling price through discounted cash flow analysis, were calculated using an Aspen Plus process simulation with techno-economic analysis (TEA) [@humbird2011process]. 
+`VE` was originally developed to support the simulation and optimization of the biochemical conversion of lignocellulosic biomass to fuels. This bioconversion process was modeled by linking previously developed computational models as three important unit operations (Figure \ref{fig:VE_diagram}):  (i) the _pretreatment_ of the feedstock to make cellulose more accessible [@sitaraman_multiphysics_2015], (ii) an _enzymatic hydrolysis_ step to digest lingnocellulose into sugars [@sitaraman_coupled_2019;@lischeske2019two], and (iii) a _bioconversion_ step to convert sugars into products in a bioreactor [@rahimi_computational_2018].  Finally, the capital and operating costs of the process, and the product's (in this case, ethanol) subsequent minimum feasible selling price through discounted cash flow analysis, were calculated using an Aspen Plus process simulation with techno-economic analysis (TEA) [@humbird2011process]. 
 
 ![An example of the end-to-end process for biochemical convertion of biomass to fuel defined and automated within `VE` by connecting different unit models.\label{fig:VE_diagram}](figs/VE_diagram.pdf){ width=100% }
 
@@ -87,9 +92,9 @@ We solve this optimization problem for two different modeled feedstocks: switchg
 
[…]
