### 2025-03-23 #7498 @prashjha
> Hi [@prashjha](https://github.com/prashjha) and thank you for considering me for this one. I did scan through the repo + documentation a bit and would have a couple of questions before taking the review. [@olgadoronina](https://github.com/olgadoronina) might be able to answers them:
>
> * I understood, that `VirtualEngineering` controls the different unit operation simulations by allowing users to input information and then automatically executing the individual models by passing the input information as well as upstream model results. Furthermore, automatic workflows for optimization and parametric studies are possible. This is the main feature of `VirtualEngineering`?
> * What I am not sure about:
>
>   * How are the models connected exactly, what controls the passing of parameters? There is a template model available, which seems straight forward, but there does not seem to be any documentation on what the best practices are to make your custom models talk to each other in the online documentation?
>   * What would then be the difference to other workflow management tools, e.g. [snakemake](https://snakemake.readthedocs.io/), except there is the availability of graphical user input/output through notebooks?
>
> Thank you and have a nice weekend!
>
> Francesco

Hi @olgadoronina, Could you respond to @fwitte's comments so that he can decide whether to accept the invitation to review or not?
