### 2025-03-23 #7498 @prashjha
> Thank you, [@prashjha](https://github.com/prashjha), we've edited the markdown file to reduce the word count closer to 1500 (not including references and acknowledgements). Going much below this would require additional cuts to our examples that might diminish the overall message, but let us know if more is needed. Thank you for all your help!
>
> Ethan

Hi Ethan, okay, we leave this as it is for now. It should be okay.
