### 2025-04-14 #8050 @editorialbot
[REVIEW]: Virtual Engineering: Python framework for engineering process design
**Submitting author:** @olgadoronina (<a href="http://orcid.org/0000-0003-3413-3197">Olga Doronina</a>)
**Repository:** https://github.com/NREL/VirtualEngineering
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
