### 2025-04-16 #8050 @mustafaalsalmi1999
@prashjha the installation instructions can be improved by navigating to the main "Virtual Engineering" directory
