### 2025-05-12 f13ed56
Small changes to the installation instructions
