+++ README.md
@@ -16,7 +16,7 @@ First, create a Conda environment from the `environment.yaml` file by opening a 
 
 `conda env create -f environment.yaml -n [environment name]`
 
-where `[environment name]` is replaced with the desired name of your VE Conda environment. This Conda installs the required packages and pip installs the `vebio` Python package.  Next, activate the environment with
+where `[environment name]` is replaced with the desired name of your VE Conda environment. This Conda installs the required packages and pip installs the `virteng` Python package.  Next, activate the environment with
 
 `conda activate [environment name]`
 
@@ -24,11 +24,11 @@ and launch the Notebook by running
 
 `jupyter notebook`
 
-When the Notebook landing page opens, you'll be able to open the [virtual_engineering_notebook](virtual_engineering_notebook.md) which will guide you through the next steps.
+When the Notebook landing page opens, you'll be able to navigate to the ``application/Template_aplication`` directory and open the ``virtual_engineering_notebook_template.ipynb`` which will guide you through the next steps.
 
-## Submodules
+## Git Submodules
 
-Git has a mechanism for code in a repository to depend on code in another repository. The dependency is known as a [submodule](https://git-scm.com/book/en/v2/Git-Tools-Submodules). Submodules are specified in `.gitmodules` (a hidden file on some systems) and are located in subfolders of `submodules/` of this repository. Initiate the submodule system with
+Some applications (e.g. Bioconversion application) might depend on code in another Git repositories. The dependency is known as a [submodule](https://git-scm.com/book/en/v2/Git-Tools-Submodules). Submodules are specified in `.gitmodules` (a hidden file on some systems) and are located in subfolders of `submodules/` of application repository. Initiate the submodule system with
 
 `git submodule init`
 
@@ -36,5 +36,5 @@ Then update specific modules with
 
 `git submodule update submodules/[name of submodule]`
 
-This command will perform a git clone or pull for that module. To run the enzymatic hydrolysis CFD model, update `Nek5000`. To run the well-mixed "Ligncellulose" model (no CFD), update`CEH_EmpiricalModel`. To run bioreactor CFD simulation, update `OpenFOAM-dev` and `ThirdParty-dev`. Also see [README_OpenFOAM.md](README_OpenFOAM.md).
+This command will perform a git clone or pull for that module. For example, in Bioconversion application, to run the enzymatic hydrolysis CFD model, update `Nek5000`. To run the well-mixed "Ligncellulose" model (no CFD), update`CEH_EmpiricalModel`. To run bioreactor CFD simulation, update `OpenFOAM-dev` and `ThirdParty-dev`. For example, see [README_OpenFOAM](https://virtualengineering.readthedocs.io/en/latest/applications/bioconversion/build_openfoam.html).
 
