### 2025-07-31 #8050 @mustafaalsalmi1999
> Hi [@mustafaalsalmi1999](https://github.com/mustafaalsalmi1999), we have a sentence guiding users to navigate to the root level of the Virtual Engineering directory before creating a conda environment:
>
> > First, create a Conda environment from the `environment.yaml` file by opening a terminal, navigating to the root level of the Virtual Engineering directory, and running:
>
> Please let us know if there is something else we are missing

Clear
