### 2025-08-04 #8050 @victoraalves
@olgadoronina Overall a nice contribution. I'm mostly done with my review and tests. A few comments:

1. Maybe having automatic package installation would be a nice addition (e.g., pip/conda)
2. Is there any place in the documentation that shows clear instructions on how to (a) contribute to the package (b) seek support and (c) report issues? Maybe I missed it. If not, it would a good addition as well to the documentation/readme.md file.
