+++ .github/workflows/test_vebio.yml
@@ -27,7 +27,7 @@ jobs:
             os-version: macos-latest
     steps:
     - uses: actions/checkout@v3
-    - uses: conda-incubator/setup-miniconda@v4
+    - uses: conda-incubator/setup-miniconda@v2
       with:
         miniconda-version: "latest"
         channels: conda-forge
