### 2025-08-25 cf56eca
Added installation instruction
