### 2025-08-25 67284dc
minot change to installation instructions
