### 2025-09-17 #8050 @victoraalves
> [@victoraalves](https://github.com/victoraalves) We added the `virtualengineering` package to conda-forge and added installation instructions to README.md. We also added a reference to CONTRIBUTING.md at the end of README.md

LGTM!
