### 2025-09-28 #8050 @prashjha
Hi @olgadoronina, could you please archive (if not already done) the release using Zenodo and provide the archive reference so that I can associate it with your JOSS submission? Also, please make sure that the title of the Zenodo archive matches the title of this JOSS article.

If you have updated a version of your code, let me know, and I can update it here.

Below are my suggestions on your draft:
- L11: 'joins them to simulate an entire end-to-end process' -> 'integrates them to simulate a complete end-to-end process'?
- Fig 1 caption: 'convertion' -> 'conversion'
- L104: 'analaysis' -> 'analysis'
- L114: 'provide overall optimal process conditions' -> 'identify overall optimal process conditions'
