### 2025-09-29 18b5011
Fix typos and improve clarity in paper.md
