+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 # Summary
 
 `Virtual Engineering` (`VE`) is a Python software framework designed to accelerate the research and development of engineering processes that are fundamentally defined by multiple unit operations executed in series. 
-VE supports a wide variety of different multi-physics models and joins them to simulate an entire end-to-end process. To automate the execution of this model sequence, `VE` provides (i) a robust method to communicate between models, 
+VE supports a wide variety of different multi-physics models and integrates them to simulate a complete end-to-end process. To automate the execution of this model sequence, `VE` provides (i) a robust method to communicate between models, 
 (ii) a high-level, user-friendly interface to set model parameters and enable optimization, and (iii) an overall model-agnostic approach that allows new computational units to be swapped in and out of workflows. 
 Although the `VE` framework was developed to support the biochemical conversion of biomass to fuel, we have designed each component to easily accommodate new domains and unit models.
 
@@ -45,7 +45,7 @@ Numerical simulations that can support the analysis and optimization of such sys
 
 `VE` was originally developed to support the simulation and optimization of the biochemical conversion of lignocellulosic biomass to fuels. This bioconversion process was modeled by linking previously developed computational models as three important unit operations (Figure \ref{fig:VE_diagram}):  (i) the _pretreatment_ of the feedstock to make cellulose more accessible [@sitaraman_multiphysics_2015], (ii) an _enzymatic hydrolysis_ step to digest lignocellulose into sugars [@sitaraman_coupled_2019;@lischeske2019two], and (iii) a _bioconversion_ step to convert sugars into products in a bioreactor [@rahimi_computational_2018].  Finally, the capital and operating costs of the process, and the product's (in this case, ethanol) subsequent minimum feasible selling price through discounted cash flow analysis, were calculated using an Aspen Plus process simulation with techno-economic analysis (TEA) [@humbird2011process]. 
 
-![An example of the end-to-end process for biochemical convertion of biomass to fuel defined and automated within `VE` by connecting different unit models.\label{fig:VE_diagram}](figs/VE_diagram.pdf){ width=100% }
+![An example of the end-to-end process for biochemical conversion of biomass to fuel defined and automated within `VE` by connecting different unit models.\label{fig:VE_diagram}](figs/VE_diagram.pdf){ width=100% }
 
 ![Example of `ipywidgets` element defining the choices of EH unit model and its input parameters.\label{fig:controls}](figs/eh_widgets.png){ width=100% }
 
@@ -64,7 +64,7 @@ To enable users to easily set up and launch new simulations, either by swapping 
 
 The `VE` package provides an optimization capability that can be easily accessed through the user-friendly `ipywidgets` GUI interface. The notebook supports the selection of multiple control variables, in which case all chosen variables will be tuned to minimize a given objective function over a multidimensional design space. Additionally, users can choose to minimize either the final output or any intermediate output of the process. In the latter case, the number of unit models in the process is automatically adjusted based on the specified objective; for example, executing only the first element in a three-element series if the objective function depends only on element one.
 
-To demonstrate these capabilities of the `VE` package, we optimize the oxygen uptake rate (OUR) that determine fuel yields in the bioconversion process (Figure \ref{fig:VE_diagram}), by solving the box-constrained optimization problem with two controls summarized in Table \ref{tab:opt_problem}. It is important to note that the objective function, OUR, is an output of the bioreactor (BR) model, while acid loading and enzyme loading are independent inputs to the pretreatment (PT) and enzymatic hydrolysis (EH) models, respectively, which indirectly affect overall fuel yields. 
[…]
