### 2025-09-29 b3490ef
Delete paper_old.md
