### 2025-09-29 #8050 @olgadoronina
Hi @prashjha, I have updated the paper with your suggestions and created a new release (v1.0.0), which I have archived on Zenodo. Here is the DOI: https://doi.org/10.5281/zenodo.17227544
